initial commit of UGNT
Finished Matt-Rev
Checked all UB WH UBS4 for { }
Where there was a { }, I manually chose the reading found in CNTR's
majority, earliest
This is essential because I see now that where UB WH UBS4 has { } it has
NOT including the Greek text. In other words, it leaves out the Greek
text which is listed within the { } so it is a must that we check all of
them before releasing the text. Otherwise we know we are missing text.
Also, it is a fast way to only check variants which are differences
between WH & UBS4.